requirements.txt
requirements.txt => fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
openai==1.12.0
google-generativeai==0.3.2